added copying history

+++ Cabinet.py
@@ -32,14 +32,14 @@ def image_information(image_path):
         except AttributeError:
             return({"coordinates":False, "year":img.datetime_original.split(":")[0]})
     else:
-        raise Exception('The Image has no EXIF information')
+        raise Exception("The Image has no EXIF information")
         
     return({"coordinates":True, "year":img.datetime_original.split(":")[0], "latitude":coords[0],"longitude":coords[1]})
 
 # Makes the path to which copy the image based on its metadata
 def make_path(info):
     if info["coordinates"]:
-        location = geoloc.reverse(f"{str(info['latitude'])},{str(info['longitude'])}", language="en").raw['address']
+        location = geoloc.reverse(f"{str(info['latitude'])},{str(info['longitude'])}", language="en").raw["address"]
 
         country = location['country']
 
@@ -74,6 +74,8 @@ def make_path(info):
     return path
 
 def main():
+    copied = 0
+
     for file in os.listdir(image_dir):
         if file.endswith(".jpg"):
 
@@ -84,15 +86,33 @@ def main():
 
             # If there's no EXIF metadata available, copy it to the unknown subfolder
             except:
-                print(f"Error. Image {file} couldn't be classified. Moving to {destination_dir}/Unknown/")
-                os.makedirs(os.path.dirname(f"{destination_dir}/Unknown/"), exist_ok=True)
-                shutil.copyfile(f"{image_dir}/{file}", f"{destination_dir}/Unknown/{file}")
+
+                #Verify if an image with the same name exists
+                if(os.path.isfile(f"{destination_dir}/Unknown/")):
+                    print(f"Image {file} is already in destination directory. Skipping")
+                else:
+                    print(f"Error. Image {file} couldn't be classified. Moving to {destination_dir}/Unknown/")
+                    os.makedirs(os.path.dirname(f"{destination_dir}/Unknown/"), exist_ok=True)
+                    shutil.copyfile(f"{image_dir}/{file}", f"{destination_dir}/Unknown/{file}")
+
+                    copied += 1
             
-            # 
+            # If there was metadata, copy the image to the generated path, first verify if an image with the same name exists
+            # Which are the odds of two different images having the exact same name and location? Skip image if so
             else:
-                print(path)
-                os.makedirs(os.path.dirname(path + '/'), exist_ok=True)
-                shutil.copyfile(f"{image_dir}/{file}", f"{path}/{file}")
+                if(os.path.isfile(f"{path}/{file}")):
+                    print(f"Image {file} is already in destination directory. Skipping")
+                else:
+                    print(path)
+                    os.makedirs(os.path.dirname(path + '/'), exist_ok=True)
+                    shutil.copyfile(f"{image_dir}/{file}", f"{path}/{file}")
+
+                    copied += 1
+
+    # Write the copying date into the record, for reference purpose when making future copyings
+    file = open(f"{destination_dir}/CabinetHistory.txt", "a")
+    record = f"{datetime.now()}, {copied} images copied"
+    file.write()
 
 if __name__ == '__main__':
     main()
